branches/saulius-generic-types/ (saulius@kolibris) Aligning generic types in structures at 'int' boundary, not at 'float' boundary.
git-svn-id: svn+ssh://saulius.grazulis.lt/home/saulius/svn-repositories/compilers/branches/saulius-generic-types@3361 f7c8cbe2-f9e6-0310-9050-d7f47bf15973

+++ sc/interpreters/packed-type/tnode.c
@@ -1236,7 +1236,7 @@ TNODE *tnode_insert_fields( TNODE* tnode, DNODE *field )
             ssize_t old_offset = tnode->nextnumoffs;
             field_size = sizeof(union stackunion);
 #if 1
-            field_align = sizeof(float);
+            field_align = sizeof(int);
             ALIGN_NUMBER( tnode->nextnumoffs, field_align );
 #endif
             /* printf( ">>> bits = %d, max_size = %d\n", bits, max_size ); */
